# src/main/java/com/parko/balance/service/dto/request/ChargeRequest.java
package com.parko.balance.service.dto.request;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;

import java.math.BigDecimal;
import java.util.UUID;

public record ChargeRequest(
        @NotNull UUID userId,
        @NotNull @Positive BigDecimal amount
) {
}